move humanoid specific VRML writer code under src/Body
