+++ include/cnoid/VRMLBody
@@ -1,0 +1,1 @@
+#include "src/Body/VRMLBody.h"
