+++ include/cnoid/VRMLBodyWriter
@@ -1,0 +1,1 @@
+#include "src/Body/VRMLBodyWriter.h"
